# priv/static/swagger.json
{
  "definitions": {
...
    "CreateWordRequest": {
      "description": "POST body for creating a word",
      "example": {
        "word": {
          "description": "The Operation in ISO 9001 2015 Clause 8 of the 2015 revision of the ISO 9001 standard seeks to improve the operational control of your organization's production processes. In short, it asks you to effectively define the criteria and processes for the products and services to be delivered to customers, as well as the documentation and resources to ensure that the products and services you offer are adequate.",
          "name": "Operation"
        }
      },
      "properties": {
        "word": {
          "$ref": "#/definitions/Word",
          "description": "The word details"
        }
      },
      "title": "CreateWordRequest",
      "type": "object"
    },
    "CreateWordResponse": {
      "description": "Response schema for single word",
      "example": {
        "data": {
          "description": "The Operation in ISO 9001 2015 Clause 8 of the 2015 revision of the ISO 9001 standard seeks to improve the operational control of your organization's production processes. In short, it asks you to effectively define the criteria and processes for the products and services to be delivered to customers, as well as the documentation and resources to ensure that the products and services you offer are adequate.",
          "id": "e9bd02f8-f81a-4e4b-b67c-86df37b4ff1a",
          "image": {
            "file_name": "operation.png",
            "updated_at": "2022-12-16T15:55:28"
          },
          "name": "Operation"
        }
      },
      "properties": {
        "data": {
          "$ref": "#/definitions/Word",
          "description": "The word details created"
        }
      },
      "title": "CreateWordResponse",
      "type": "object"
    },
    "CreateWordResponseErrors": {
      "description": "Response errors from create word",
      "example": {
        "description": "can't be blank",
        "name": "can't be blank"
      },
      "title": "CreateWordResponseErrors",
      "type": "object"
    },
...
    "Error": {
      "description": "Error responses from the API",
      "properties": {
        "error": {
          "description": "The message of the error raised",
          "type": "string"
        }
      },
      "required": [
        "error"
      ],
      "title": "Errors",
      "type": "object"
    },
...
    "ShowWordsEmptyResponse": {
      "description": "Response error words empty",
      "example": {
        "error": "There are no words"
      },
      "title": "ShowWordsEmptyResponse",
      "type": "object"
    },
    "ShowWordsResponse": {
      "description": "Response schema for multiple words created",
      "example": {
        "first": 1,
        "has_next": false,
        "has_prev": false,
        "last": 2,
        "next_page": 2,
        "page": 1,
        "prev_page": 0,
        "total_count": 2,
        "words": [
          {
            "category_id": "5a10f256-51cd-4e91-844f-13d17a65ad57",
            "description": "The Operation in ISO 9001 2015 Clause 8 of the 2015 revision of the ISO 9001 standard seeks to improve the operational control of your organization's production processes. In short, it asks you to effectively define the criteria and processes for the products and services to be delivered to customers, as well as the documentation and resources to ensure that the products and services you offer are adequate.",
            "id": "e9bd02f8-f81a-4e4b-b67c-86df37b4ff1a",
            "image": {
              "file_name": "operation.png",
              "updated_at": "2022-12-16T15:55:28"
            },
            "name": "Operation"
          },
          {
            "category_id": "5a10f256-51cd-4e91-844f-13d17a65ad57",
            "description": "For the ISO-9000 standard, quality is the degree to which a set of inherent characteristics meets the requirements, being: Quality characteristic: inherent characteristic of a product, process or system related to a requirement.",
            "id": "df48a064-75d5-4be7-a773-7003169277e6",
            "image": {
              "file_name": "quality.png",
              "updated_at": "2022-12-16T15:55:28"
            },
            "name": "Quality"
          }
        ]
      },
      "properties": {
        "words": {
          "description": "The words details",
          "items": {
            "$ref": "#/definitions/Word"
          },
          "type": "array"
        }
      },
      "title": "ShowWordsResponse",
      "type": "object"
    },
...
    "UpdateImageWordRequest": {
      "description": "Request params to update a word's image",
      "example": {
        "image": "operation.png"
      },
      "properties": {
        "data": {
          "description": "The words params",
          "items": {
            "$ref": "#/definitions/Word"
          },
          "type": "array"
        }
      },
      "title": "UpdateImageWordRequest",
      "type": "object"
    },
...
    "UpdateWordRequest": {
      "description": "Request params to update a word",
      "example": {
        "name": "Authority to dispose of"
      },
      "properties": {
        "data": {
          "description": "The word params",
          "items": {
            "$ref": "#/definitions/Word"
          },
          "type": "array"
        }
      },
      "title": "UpdateWordRequest",
      "type": "object"
    },
    "UpdateWordResponse": {
      "description": "Response schema of a single word",
      "example": {
        "data": {
          "description": "The Operation in ISO 9001 2015 Clause 8 of the 2015 revision of the ISO 9001 standard seeks to improve the operational control of your organization's production processes. In short, it asks you to effectively define the criteria and processes for the products and services to be delivered to customers, as well as the documentation and resources to ensure that the products and services you offer are adequate.",
          "id": "e9bd02f8-f81a-4e4b-b67c-86df37b4ff1a",
          "image": {
            "file_name": "operation.png",
            "updated_at": "2022-12-16T15:55:28"
          },
          "name": "Authority to dispose of"
        }
      },
      "properties": {
        "data": {
          "description": "The word updated",
          "items": {
            "$ref": "#/definitions/Word"
          },
          "type": "array"
        }
      },
      "title": "UpdateWordResponse",
      "type": "object"
    },
...
  },
  "host": "localhost:4000",
  "info": {
    "contact": {
      "email": "orlando@cordage.io",
      "name": "Orlando Gomez"
    },
    "description": "API Documentation for GreenHouse API v1",
    "termsOfService": "Open for public",
    "title": "Words GreenHouse API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/categories/{category_id}/words": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "deprecated": false,
        "description": "Create a new word in the database",
        "operationId": "GlossaryWeb.WordController.create",
        "parameters": [
          {
            "description": "The ID of the category",
            "in": "path",
            "name": "category_id",
            "required": true,
            "type": "string",
            "x-example": "1b1716bd-1875-44c4-8c4b-6f969fc5fdbd"
          },
          {
            "description": "The word data",
            "in": "body",
            "name": "word",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateWordRequest"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/CreateWordResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/CreateWordResponseErrors"
            }
          }
        },
        "summary": "Add a new word",
        "tags": [
          "Word"
        ]
      }
    },
    "/api/v1/categories/{category_id}/words/image/{image}": {
      "get": {
        "consumes": [
          "multipart/form-data"
        ],
        "description": "Show the word's description picture",
        "operationId": "GlossaryWeb.WordController.show_image",
        "parameters": [
          {
            "description": "The ID of the category",
            "in": "path",
            "name": "category_id",
            "required": true,
            "type": "string",
            "x-example": "1b1716bd-1875-44c4-8c4b-6f969fc5fdbd"
          },
          {
            "description": "Word ID",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "x-example": "1b1716bd-1875-44c4-8c4b-6f969fc5fdbd"
          }
        ],
        "produces": [
          "multipart/form-data"
        ],
        "responses": {
          "201": {
            "description": "Ok",
            "schema": {
              "$ref": "#/definitions/UpdateImageWordRequest"
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "summary": "Show the image of a word",
        "tags": [
          "Word"
        ]
      }
    },
    "/api/v1/categories/{category_id}/words/search?page={page}&page_size={size}&name={name}": {
      "get": {
        "deprecated": false,
        "description": "Returns JSON with all words matching listed in the database",
        "operationId": "GlossaryWeb.WordController.search_word",
        "parameters": [
          {
            "description": "Page number to be displayed",
            "in": "path",
            "name": "page",
            "required": true,
            "type": "integer",
            "x-example": 2
          },
          {
            "description": "Total number of entries per page",
            "in": "path",
            "name": "size",
            "required": true,
            "type": "integer",
            "x-example": 10
          },
          {
            "description": "The field you want to find",
            "in": "path",
            "name": "name",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ShowWordsResponse"
            }
          },
          "204": {
            "description": "No users",
            "schema": {
              "$ref": "#/definitions/ShowWordsEmptyResponse"
            }
          }
        },
        "summary": "All Words in the search",
        "tags": [
          "Word"
        ]
      }
    },
    "/api/v1/categories/{category_id}/words/update/{id}": {
      "delete": {
        "description": "Delete a word by ID",
        "operationId": "GlossaryWeb.WordController.delete",
        "parameters": [
          {
            "description": "The ID of the category",
            "in": "path",
            "name": "category_id",
            "required": true,
            "type": "string",
            "x-example": "1b1716bd-1875-44c4-8c4b-6f969fc5fdbd"
          },
          {
            "description": "Word ID",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "x-example": "1b1716bd-1875-44c4-8c4b-6f969fc5fdbd"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content - Deleted Successfully"
          }
        },
        "summary": "Delete a specific word",
        "tags": [
          "Word"
        ]
      },
      "put": {
        "consumes": [
          "application/json"
        ],
        "description": "Update all attributes of a word",
        "operationId": "GlossaryWeb.WordController.update",
        "parameters": [
          {
            "description": "The ID of the category",
            "in": "path",
            "name": "category_id",
            "required": true,
            "type": "string",
            "x-example": "1b1716bd-1875-44c4-8c4b-6f969fc5fdbd"
          },
          {
            "description": "Word ID",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "x-example": "1b1716bd-1875-44c4-8c4b-6f969fc5fdbd"
          },
          {
            "description": "The word details to update",
            "in": "body",
            "name": "word",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateWordRequest"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "205": {
            "description": "Updated",
            "schema": {
              "$ref": "#/definitions/UpdateWordResponse"
            }
          },
          "400": {
            "description": "Unprocessable Entity",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "summary": "Update a word",
        "tags": [
          "Word"
        ]
      }
    },
    "/api/v1/categories/{category_id}/words/upload_image/{id}": {
      "put": {
        "consumes": [
          "multipart/form-data"
        ],
        "description": "Update the word's description picture",
        "operationId": "GlossaryWeb.WordController.upload_image",
        "parameters": [
          {
            "description": "The ID of the category",
            "in": "path",
            "name": "category_id",
            "required": true,
            "type": "string",
            "x-example": "1b1716bd-1875-44c4-8c4b-6f969fc5fdbd"
          },
          {
            "description": "Word ID",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "x-example": "1b1716bd-1875-44c4-8c4b-6f969fc5fdbd"
          },
          {
            "description": "Browse File",
            "in": "file",
            "name": "image",
            "required": true,
            "type": "formData"
          }
        ],
        "produces": [
          "multipart/form-data"
        ],
        "responses": {
          "201": {
            "description": "Ok",
            "schema": {
              "$ref": "#/definitions/UpdateImageWordRequest"
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "summary": "Upload image of a word",
        "tags": [
          "Word"
        ]
      }
    },
    "/api/v1/categories/{category_id}/words/{id}": {
      "get": {
        "deprecated": false,
        "description": "Return JSON with an especific word",
        "operationId": "GlossaryWeb.WordController.show",
        "parameters": [
          {
            "description": "The ID of the category",
            "in": "path",
            "name": "category_id",
            "required": true,
            "type": "string",
            "x-example": "1b1716bd-1875-44c4-8c4b-6f969fc5fdbd"
          },
          {
            "description": "Word ID",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "x-example": "1b1716bd-1875-44c4-8c4b-6f969fc5fdbd"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Word created OK",
            "schema": {
              "$ref": "#/definitions/ShowWordResponse"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "summary": "Show a Specific Word",
        "tags": [
          "Word"
        ]
      }
    },
    "/api/v1/categories/{category_id}/words?page={page}&page_size={size}": {
      "get": {
        "deprecated": false,
        "description": "Return JSON with all words listed in the database",
        "operationId": "GlossaryWeb.WordController.index",
        "parameters": [
          {
            "description": "The ID of the category",
            "in": "path",
            "name": "category_id",
            "required": true,
            "type": "string",
            "x-example": "1b1716bd-1875-44c4-8c4b-6f969fc5fdbd"
          },
          {
            "description": "Page number to be displayed",
            "in": "path",
            "name": "page",
            "required": true,
            "type": "integer",
            "x-example": 2
          },
          {
            "description": "Total number of entries per page",
            "in": "path",
            "name": "size",
            "required": true,
            "type": "integer",
            "x-example": 10
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ShowWordsResponse"
            }
          },
          "204": {
            "description": "No words",
            "schema": {
              "$ref": "#/definitions/ShowWordsEmptyResponse"
            }
          }
        },
        "summary": "All Words",
        "tags": [
          "Word"
        ]
      }
    },
...
  },
  "swagger": "2.0"
}